extract the GTM consent composer as standalone
